rename select token modal file and move to components
